requirements.txt
django
djangorestframework
drf-compound-fields
